Refactor architecture of ocr backend
