+++ webapp/flaskr/ocr/paddle_hw_reader.py
@@ -1,5 +1,5 @@
 from paddleocr import PaddleOCR
-from handwriting_reader import HandwritingReader, ReadText
+from ocr.abstract_hw_reader import HandwritingReader, ReadText
 import logging
 from pathlib import Path
 
@@ -11,7 +11,6 @@ class PaddleHandwritingReader(HandwritingReader):
         self.client = PaddleOCR(
             text_detection_model_name="PP-OCRv5_server_det",
             text_recognition_model_name="en_PP-OCRv5_mobile_rec",
-            lang="en",
             use_doc_orientation_classify=False,
             use_doc_unwarping=False,
             use_textline_orientation=False,
